FirebaseConnect cleanup
