+++ database.rules.json
@@ -1,6 +1,6 @@
 {
   "rules": {
-    ".read": "auth != null",
-    ".write": "auth != null"
+    ".read": true
+    ".write": true
   }
 }
